Fixed download jobs rest urls and added javadoc

+++ goobi-viewer-core/src/de/intranda/digiverso/presentation/servlets/rest/download/DownloadResource.java
@@ -33,6 +33,7 @@ import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 
 import org.apache.commons.lang.StringEscapeUtils;
+import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -57,12 +58,25 @@ public class DownloadResource {
     @Context
     private HttpServletRequest servletRequest;
 
+    /**
+     * Get information about a specific downloadJob
+     * 
+     * @param type  The jobtype, either pdf or epub
+     * @param pi    The PI of the underlying record
+     * @param logId The logId of the underyling docStruct. Is ignored if it matches the regex [-(null)]/i
+     * @return  A json representation of the {@link DownloadJob}
+     * @throws DAOException     If an error occured querying the database
+     * @throws ContentNotFoundException If no matching job was found
+     */
     @GET
     @Path("/get/{type}/{pi}/{logId}")
     @Produces({ MediaType.APPLICATION_JSON })
     @AuthenticationBinding
     public DownloadJob getDownloadInfo(@PathParam("type") String type, @PathParam("pi") String pi, @PathParam("logId") String logId)
             throws DAOException, ContentNotFoundException {
+        if(StringUtils.isBlank(logId.replaceAll("(?i)[-(null)]", ""))) {
+            logId = null;
+        }
         DownloadJob downloadJob = DataManager.getInstance().getDao().getDownloadJobByMetadata(type, pi, logId);
         if (downloadJob != null) {
             return downloadJob;
@@ -71,20 +85,36 @@ public class DownloadResource {
         }
     }
     
+    /**
+     * Get information about a specific downloadJob
+     * 
+     * @param type  The jobtype, either pdf or epub
+     * @param identifier    The job idenfier
+     * @return  A json representation of the {@link DownloadJob}
+     * @throws DAOException     If an error occured querying the database
+     * @throws ContentNotFoundException If no matching job was found
+     */
     @GET
-    @Path("/get/{identifier}")
+    @Path("/get/{type}/{identifier}")
     @Produces({ MediaType.APPLICATION_JSON })
     @AuthenticationBinding
-    public DownloadJob getDownloadInfo(@PathParam("identifier") String identifier)
+    public DownloadJob getDownloadInfo(@PathParam("identifier") String identifier, @PathParam("type") String type)
             throws DAOException, ContentNotFoundException {
         DownloadJob downloadJob = DataManager.getInstance().getDao().getDownloadJobByIdentifier(identifier);
-        if (downloadJob != null) {
+        if (downloadJob != null && downloadJob.getType().equalsIgnoreCase(type)) {
             return downloadJob;
         } else {
-            throw new ContentNotFoundException("No download job found for identifier " + identifier);
+            throw new ContentNotFoundException("No " + type + " download job found for identifier " + identifier);
         }
     }
     
+    /**
+     * Get information about all download jobs of a type
+     * 
+     * @param type  The jobtype, either pdf or epub
+     * @return  An array of json representations of all {@link DownloadJob}s of the given type
+     * @throws DAOException     If an error occured querying the database
+     */
     @GET
     @Path("/get/{type}")
     @Produces({ MediaType.APPLICATION_JSON })
@@ -99,12 +129,25 @@ public class DownloadResource {
         return downloadJobs;
     }
 
+    /**
+     * Remove a download job from the database
+     * 
+     * @param type  The jobtype, either pdf or epub
+     * @param pi    The PI of the underlying record
+     * @param logId The logId of the underyling docStruct. Is ignored if it matches the regex [-(null)]/i
+     * @return  A json object containing the job identifier and wether the job could be deleted
+     * @throws DAOException     If an error occured querying the database
+     * @throws ContentNotFoundException If no matching job was found
+     */
     @GET
     @Path("/delete/{type}/{pi}/{logId}")
     @Produces({ MediaType.APPLICATION_JSON })
     @AuthenticationBinding
     public String deleteDownloadJob(@PathParam("type") String type, @PathParam("pi") String pi, @PathParam("logId") String logId)
             throws DAOException, ContentLibException {
+        if(StringUtils.isBlank(logId.replaceAll("(?i)[-(null)]", ""))) {
+            logId = null;
+        }
         DownloadJob downloadJob = DataManager.getInstance().getDao().getDownloadJobByMetadata(type, pi, logId);
         if (downloadJob != null) {
             if (!DataManager.getInstance().getDao().deleteDownloadJob(downloadJob)) {
@@ -117,24 +160,41 @@ public class DownloadResource {
         }
     }
     
+    /**
+     * Remove a download job from the database
+     * 
+     * @param type  The jobtype, either pdf or epub
+     * @param identifier    The job idenfier
+     * @return  A json object containing the job identifier and wether the job could be deleted
+     * @throws DAOException     If an error occured querying the database
+     * @throws ContentNotFoundException If no matching jobs were found
+     */
     @GET
-    @Path("/delete/{identifier}")
+    @Path("/delete/{type}/{identifier}")
     @Produces({ MediaType.APPLICATION_JSON })
     @AuthenticationBinding
-    public String deleteDownloadJob(@PathParam("identifier") String identifier)
+    public String deleteDownloadJob(@PathParam("type") String type, @PathParam("identifier") String identifier)
             throws DAOException, ContentLibException {
         DownloadJob downloadJob = DataManager.getInstance().getDao().getDownloadJobByIdentifier(identifier);
-        if (downloadJob != null) {
+        if (downloadJob != null && downloadJob.getType().equalsIgnoreCase(type)) {
             if (!DataManager.getInstance().getDao().deleteDownloadJob(downloadJob)) {
                 return "{job: \"" + downloadJob.getIdentifier() + "\", deleted: false}";
             } else {
                 return "{job: \"" + downloadJob.getIdentifier() + "\", deleted: true}";
             }
         } else {
-            throw new ContentNotFoundException("No download job found for identifier " + identifier);
+            throw new ContentNotFoundException("No " + type + " download job found for identifier " + identifier);
         }
     }
 
+    /**
+     * Remove all jobs of a type from the database
+     * 
+     * @param type  The jobtype, either pdf or epub
+     * @return  An array of json objects containing the job identifiers and wether the jobs could be deleted
+     * @throws DAOException     If an error occured querying the database
+     * @throws ContentNotFoundException If no matching job was found
+     */
     @GET
     @Path("/delete/{type}")
     @Produces({ MediaType.APPLICATION_JSON })
@@ -147,7 +207,7 @@ public class DownloadResource {
                 .filter(job -> type.equalsIgnoreCase("all") || job.getType().equalsIgnoreCase(type))
                 .collect(Collectors.toList());
         if (!downloadJobs.isEmpty()) {
-            List<String> results = new ArrayList();
+            List<String> results = new ArrayList<>();
             for (DownloadJob job : downloadJobs) {
                 if(DataManager.getInstance().getDao().deleteDownloadJob(job)) {
                     results.add("{job: \"" + job.getIdentifier() + "\", deleted: true}");
